+++ frontend/src/components/Timer.tsx
@@ -1,35 +1,9 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
+import { useCountTime } from '../hooks/hooks';
 import TimeDisplay from './TimeDisplay';
 
 function Timer () {
-  const [isStarted, setIsStarted] = useState<boolean>(false)
-  const [startTime, setStartTime] = useState<Date>(new Date())
-  const [seconds, setSeconds] = useState<number>(0)
-
-  useEffect(() => {
-    if (isStarted) {
-      const intervalID = setInterval(() => {
-        const newSeconds = seconds - Math.floor((new Date().getTime() - startTime.getTime()) / (1000))
-        console.log(newSeconds);
-        setSeconds(newSeconds)
-        if (newSeconds === 0) {
-          setIsStarted(false)
-          clearInterval(intervalID)
-          return
-        }
-      }, 1000)
-
-      return () => {
-        clearInterval(intervalID)
-      }
-    }
-    console.log('Timer component mounted')
-  }, [isStarted])
-
-  const start = () => {
-    setIsStarted(true)
-    setStartTime(new Date())
-  }
+  const { isStarted, seconds, setSeconds, start } = useCountTime()
 
   return (
     <div>
